refactor: extract ErrorStack
