+++ package.json
@@ -115,7 +115,7 @@
     "execa": "5.1.1",
     "fast-deep-equal": "3.1.3",
     "flat": "5.0.2",
-    "framer-motion": "^6.3.3",
+    "framer-motion": "6.3.3",
     "git-url-parse": "11.6.0",
     "jsonpath-plus": "6.0.1",
     "lodash": "4.17.21",
