+++ src/components/organisms/ErrorPage/ErrorPage.styled.tsx
@@ -1,11 +1,4 @@
-import {Button} from 'antd';
-
-import {DownOutlined as DownOutlinedBase} from '@ant-design/icons';
-
-import {motion} from 'framer-motion';
 import styled from 'styled-components';
-
-import {GlobalScrollbarStyle} from '@utils/scrollbar';
 
 import Colors from '@styles/Colors';
 
@@ -31,23 +24,4 @@ export const Heading = styled.h1`
 
 export const Description = styled.p`
   margin-bottom: 0px;
-`;
-
-export const DownOutlined = styled(DownOutlinedBase)`
-  font-size: 10px;
-`;
-
-export const ErrorStack = styled.div`
-  margin-bottom: 6px;
-`;
-
-export const ErrorButton = styled(Button)`
-  width: 100%;
-  text-align: center;
-`;
-
-export const ErrorStackContent = styled(motion.section)`
-  height: 100%;
-  overflow-y: auto;
-  ${GlobalScrollbarStyle}
 `;
